[ALPS03695698] PMIC: Fix Coverity
[Detail]
Fix Unintentional integer overflow

MTK-Commit-Id: ed56027bea896a0671680a4c366da0bf339b9a42

CR-Id: ALPS03695698
Feature: Power Management

+++ drivers/misc/mediatek/pmic/mt6358/v1/pmic_auxadc.c
@@ -108,7 +108,7 @@ static int wk_aux_cali(int T_curr, int vbat_out)
 	signed long long coeff_gain_aux = 0;
 	signed long long vbat_cali = 0;
 
-	coeff_gain_aux = (317220 + 11960 * g_GAIN_AUX);
+	coeff_gain_aux = (317220 + 11960 * (signed long long)g_GAIN_AUX);
 	vbat_cali = div_s64((vbat_out * (T_curr - 250) * coeff_gain_aux), 255);
 	vbat_cali = div_s64(vbat_cali, 1000000000);
 	if (g_SIGN_AUX == 0)
@@ -125,13 +125,13 @@ static int wk_bgr_cali(int T_curr, int vbat_out)
 	signed int T_H = 600 + g_TEMP_H_CALI * 25;
 
 	if (T_curr < T_L) {
-		coeff_gain_bgr = (127 + 8 * g_GAIN_BGRL);
+		coeff_gain_bgr = (127 + 8 * (signed long long)g_GAIN_BGRL);
 		if (g_SIGN_BGRL == 0)
 			vbat_out += div_s64((vbat_out * (T_curr - T_L) * coeff_gain_bgr), 127000000);
 		else
 			vbat_out -= div_s64((vbat_out * (T_curr - T_L) * coeff_gain_bgr), 127000000);
 	} else if (T_curr > T_H) {
-		coeff_gain_bgr = (127 + 8 * g_GAIN_BGRH);
+		coeff_gain_bgr = (127 + 8 * (signed long long)g_GAIN_BGRH);
 		if (g_SIGN_BGRH == 0)
 			vbat_out -= div_s64((vbat_out * (T_curr - T_H) * coeff_gain_bgr), 127000000);
 		else
